chore: resolve review changes
