+++ packages/ui/src/textarea/textarea.forward-ref.test.tsx
@@ -1,94 +1,13 @@
-import { fireEvent, render, screen } from '@testing-library/react';
+import { render } from '@testing-library/react';
 import { createRef } from 'react';
-import { describe, expect, it, vi } from 'vitest';
-import { Input } from '../input/index.js';
+import { describe, expect, it } from 'vitest';
+
 import { Textarea } from './index.js';
 
-describe('Textarea', () => {
+describe('Textarea forwardRef', () => {
 	it('forwards a ref to the underlying <textarea>', () => {
 		const ref = createRef<HTMLTextAreaElement>();
 		render(<Textarea ref={ref} testId="ta" placeholder="hi" />);
 		expect(ref.current).toBeInstanceOf(HTMLTextAreaElement);
-	});
-
-	it('passes onChange through as a standard textarea ChangeEvent', () => {
-		const onChange = vi.fn();
-		render(<Textarea onChange={onChange} testId="ta" />);
-		const el = screen.getByTestId('ta') as HTMLTextAreaElement;
-		fireEvent.change(el, { target: { value: 'hello' } });
-		expect(onChange).toHaveBeenCalledTimes(1);
-		expect(onChange.mock.calls[0][0].target.value).toBe('hello');
-		expect(el.value).toBe('hello');
-	});
-
-	it('honors the rows and placeholder props', () => {
-		render(<Textarea rows={6} placeholder="Write your feedback here..." testId="ta" />);
-		const el = screen.getByTestId('ta') as HTMLTextAreaElement;
-		expect(el.rows).toBe(6);
-		expect(el.placeholder).toBe('Write your feedback here...');
-	});
-
-	it('reflects size and status on the wrapper as data attributes', () => {
-		render(<Textarea size="small" status="error" testId="ta" />);
-		const wrapper = screen.getByTestId('ta').parentElement;
-		expect(wrapper).toHaveAttribute('data-size', 'small');
-		expect(wrapper).toHaveAttribute('data-status', 'error');
-	});
-
-	it('exposes itself as Input.TextArea (AntD-style compound)', () => {
-		expect(Input.TextArea).toBe(Textarea);
-		render(<Input.TextArea testId="ta" placeholder="via Input.TextArea" />);
-		expect(screen.getByPlaceholderText('via Input.TextArea')).toBeInTheDocument();
-	});
-
-	it('marks the wrapper for autoSize and mirrors the value', () => {
-		render(<Textarea autoSize defaultValue="abc" testId="ta" />);
-		const wrapper = screen.getByTestId('ta').parentElement;
-		expect(wrapper).toHaveAttribute('data-autosize', 'true');
-		expect(wrapper).toHaveAttribute('data-replicated-value', 'abc ');
-	});
-
-	it('does not mark autoSize when the prop is absent', () => {
-		render(<Textarea testId="ta" defaultValue="abc" />);
-		const wrapper = screen.getByTestId('ta').parentElement;
-		expect(wrapper).not.toHaveAttribute('data-autosize');
-		expect(wrapper).not.toHaveAttribute('data-replicated-value');
-	});
-
-	it('updates the replicated value as the user types (uncontrolled)', () => {
-		render(<Textarea autoSize testId="ta" />);
-		const el = screen.getByTestId('ta') as HTMLTextAreaElement;
-		const wrapper = el.parentElement!;
-		fireEvent.change(el, { target: { value: 'line one\nline two' } });
-		expect(wrapper).toHaveAttribute('data-replicated-value', 'line one\nline two ');
-	});
-
-	it('updates the replicated value from a controlled value prop', () => {
-		const { rerender } = render(
-			<Textarea autoSize value="first" onChange={() => {}} testId="ta" />
-		);
-		const wrapper = screen.getByTestId('ta').parentElement!;
-		expect(wrapper).toHaveAttribute('data-replicated-value', 'first ');
-		rerender(<Textarea autoSize value="second pass" onChange={() => {}} testId="ta" />);
-		expect(wrapper).toHaveAttribute('data-replicated-value', 'second pass ');
-	});
-
-	it('exposes maxRows on the wrapper as a CSS variable', () => {
-		render(<Textarea autoSize={{ maxRows: 6 }} testId="ta" />);
-		const wrapper = screen.getByTestId('ta').parentElement!;
-		expect(wrapper).toHaveAttribute('data-has-max', 'true');
-		expect(wrapper.style.getPropertyValue('--textarea-max-rows')).toBe('6');
-	});
-
-	it('passes minRows through to the textareas `rows` attribute', () => {
-		render(<Textarea autoSize={{ minRows: 3 }} testId="ta" />);
-		const el = screen.getByTestId('ta') as HTMLTextAreaElement;
-		expect(el.rows).toBe(3);
-	});
-
-	it('testId takes precedence over data-testid when both are set', () => {
-		render(<Textarea testId="from-testid" data-testid="from-data-testid" />);
-		expect(screen.getByTestId('from-testid')).toBeInTheDocument();
-		expect(screen.queryByTestId('from-data-testid')).not.toBeInTheDocument();
 	});
 });
